Sprites and some coding
The sprites are ready (except for the mage), some changes to every enemy class have been made, a class for the ghost and the mage has been created.
Splitter added with the letter "r", ghost with the letter "g", mage and ogre still aren't implemented in the room creation yet.
The character can kill the enemies added.