My part of the enemies finished
Changes to collision: The way the collision works for the enemies has been improved now working only when the sprites are overlaping.

More music and sound effects:
the last room now plays a different track (change in spawnroom code) and sound effects have been added for the fireball after hitting the player and the ghost after being hit.

Changes in ghost, mage, fireball and names in general:
Mostly realted to collision and player tracking, the names are only changes in the sprite creations (ex: in playerChar the sprite had as name "default sprite", I changed it to "player sprite"), child splitter name have not been changed since it's empty in my version.